Refactor user management and order handling components;
