+++ server/controllers/user.controller.js
@@ -73,25 +73,45 @@ exports.updateUser = async (req, res) => {
     
     const { name, email, role, password } = req.body;
     
-    const userFields = {};
-    if (name) userFields.name = name;
-    if (email) userFields.email = email;
-    if (role) userFields.role = role;
-    if (password && password.trim() !== '') userFields.password = password;
-    
+    // Find the user first
     let user = await User.findById(id);
     
     if (!user) {
       return res.status(404).json({ message: 'User not found' });
     }
     
-    user = await User.findByIdAndUpdate(
-      id,
-      { $set: userFields },
-      { new: true }
-    ).select('-password');
-    
-    res.json(user);
+    // Handle password update separately to ensure hashing
+    if (password && password.trim() !== '') {
+      // Update password (will be hashed by pre-save hook)
+      user.password = password;
+      
+      // Update other fields if provided
+      if (name) user.name = name;
+      if (email) user.email = email;
+      if (role) user.role = role;
+      
+      // Save to trigger the pre-save hook for password hashing
+      await user.save();
+      
+      // Return user without password
+      const userResponse = user.toObject();
+      delete userResponse.password;
+      return res.json(userResponse);
+    } else {
+      // For updates without password changes, we can use findByIdAndUpdate
+      const userFields = {};
+      if (name) userFields.name = name;
+      if (email) userFields.email = email;
+      if (role) userFields.role = role;
+      
+      user = await User.findByIdAndUpdate(
+        id,
+        { $set: userFields },
+        { new: true }
+      ).select('-password');
+      
+      return res.json(user);
+    }
   } catch (error) {
     console.error(error.message);
     res.status(500).json({ message: 'Server Error' });
